Fix str object has no decode() method error. Follow-up on #94
On OSX the bulkRead method is used. That returns a bytearray. The elif branch gets triggered, which, when using Python3, fails as the constructed str object (by calling str(data)) does not posses a decode() method in Python3 any more. I don't saw any reason why we should not handle the bytearray case similarly to the str and bytes case.