Install extensions as console_scripts
The two scripts (silhouette_multi.py and sendto_silhouette.py) are now
part of the `silhouette` python pkg and the executables are created and
installed by setuptools (see the defined console scripts in setup.py).

Furthermore the pytest related code from setup.py was removed, because
the `test` subcommand of `setup.py` is deprecated.